tried to add translation into JS, fixed 404 on form removal

+++ Controller/RemoveController.php
@@ -3,6 +3,7 @@
 namespace MakingWaves\FormMakerBundle\Controller;
 
 use eZ\Bundle\EzPublishCoreBundle\Controller;
+use MakingWaves\FormMakerBundle\Entity\FormDefinitions;
 
 /**
  * Class RemoveController
@@ -12,12 +13,18 @@ class RemoveController extends Controller
 {
     /**
      * @param $formId
-     * @return \Symfony\Component\HttpFoundation\Response
+     * @return \Symfony\Component\HttpFoundation\RedirectResponse
+     * @throws \Symfony\Component\HttpKernel\Exception\NotFoundHttpException
      */
     public function removeAction( $formId )
     {
         $entityManager = $this->getDoctrine()->getManager();
         $formDefinition = $entityManager->getRepository( 'FormMakerBundle:FormDefinitions' )->find( $formId );
+
+        if ( !( $formDefinition instanceof FormDefinitions ) ) {
+            $translator = $this->get( 'translator' );
+            throw $this->createNotFoundException( $translator->trans( 'form.not.found', array(), 'formmaker' ) );
+        }
 
         $entityManager->remove( $formDefinition );
         $entityManager->flush();
